Move test files to their respective folder
