+++ packages/frontity/src/cli/__tests__/create.cli.test.ts
@@ -1,13 +1,13 @@
 import { resolve } from "path";
 import omit from "lodash.omit";
 
-import { EventPromised } from "../utils/eventPromised";
-import * as createCmd from "../commands/create";
-import create from "../cli/create";
+import create from "../create";
+import { EventPromised } from "../../utils/eventPromised";
+import * as createCmd from "../../commands/create";
 import * as inquirer from "inquirer";
-import * as utils from "../utils";
+import * as utils from "../../utils";
 
-jest.mock("../commands/create");
+jest.mock("../../commands/create");
 jest.mock("inquirer");
 
 const mockedCreateCmd = createCmd as jest.Mocked<typeof createCmd>;
